egg dropping puzzle
This is a famous puzzle problem. Suppose there is a building with n floors, if we have m eggs, then how can we find the minimum number of drops needed to find a floor from which it is safe to drop an egg without breaking it.

There some important points to remember −

When an egg does not break from a given floor, then it will not break for any lower floor also.
If an egg breaks from a given floor, then it will break for all upper floors.
When an egg breaks, it must be discarded, otherwise we can use it again.